requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
flask-marshmallow==0.11.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
more-itertools==8.2.0
Oto==1.0.1
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.4.1
pytest-mock==2.0.0
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
virtualenv==20.0.15
wcwidth==0.1.8
Werkzeug==1.0.0
